+++ android/src/main/kotlin/com/example/flutter_system_action/flutter_system_action/FlutterSystemActionPlugin.kt
@@ -2,6 +2,7 @@ package com.example.flutter_system_action.flutter_system_action
 
 import android.content.Context
 import android.content.Intent
+import android.content.IntentFilter
 import android.hardware.camera2.CameraAccessException
 import android.hardware.camera2.CameraManager
 import android.media.AudioManager
@@ -14,6 +15,10 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
+import android.content.res.Configuration
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.BatteryManager
 
 /** FlutterSystemActionPlugin */
 class FlutterSystemActionPlugin : FlutterPlugin, MethodCallHandler {
@@ -45,6 +50,34 @@ class FlutterSystemActionPlugin : FlutterPlugin, MethodCallHandler {
             "vibrationButtonPress" -> handleVibrationPressed(1)
             "ringModeOnButtonPress" -> handleVibrationPressed(2)
             "darkModeEventButtonPressed" -> darkModeEventButtonPressed()
+            "getDeviceOrientation" -> {
+                val orientation = getDeviceOrientation()
+                result.success(orientation)
+                return
+            }
+
+            "getDeviceLocale" -> {
+                val deviceLocale = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                    getDeviceLocale()
+                } else {
+                    TODO("VERSION.SDK_INT < N")
+                }
+                result.success(deviceLocale)
+                return
+            }
+
+            "getNetworkStatus" -> {
+                val networkStatus = getNetworkStatus()
+                result.success(networkStatus)
+                return
+            }
+
+            "getBatteryChargingStatus" -> {
+                val batteryChargingStatus = getBatteryChargingStatus()
+                result.success(batteryChargingStatus)
+                return
+            }
+
             else -> result.notImplemented()
         }
         result.success(null)
@@ -166,4 +199,122 @@ class FlutterSystemActionPlugin : FlutterPlugin, MethodCallHandler {
         }
         applicationContext.startActivity(intent)
     }
+
+    private fun getDeviceOrientation(): Int {
+        val orientation = applicationContext.resources.configuration.orientation
+        val orientationString = when (orientation) {
+            Configuration.ORIENTATION_PORTRAIT -> {
+                "PORTRAIT ORIENTATION"
+            }
+
+            Configuration.ORIENTATION_LANDSCAPE -> {
+                "LANDSCAPE ORIENTATION"
+            }
+
+            else -> {
+                "Unknown ORIENTATION: $orientation"
+            }
+        }
+        println(orientationString)
+
+        return when (orientation) {
+            Configuration.ORIENTATION_PORTRAIT -> 1
+            Configuration.ORIENTATION_LANDSCAPE -> 2
+            else -> 0
+        }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.N)
+    private fun getDeviceLocale(): String {
+
+        println(applicationContext.resources.configuration.locales[0].toString())
+        return applicationContext.resources.configuration.locales[0].toString()
+    }
+
+    private fun getNetworkStatus(): Int {
+        val connectivityManager =
+            applicationContext.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+
+        val iSConnectWithWifi = 1
+        val iSConnectWithMobileData = 2
+        val iSConnectWithEthernet = 3
+        val iSConnectWithUnknown = 0
+        val noInternetConnect = -1
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            val network = connectivityManager.activeNetwork
+            val capabilities = connectivityManager.getNetworkCapabilities(network)
+            if (capabilities != null) {
+                val status = when {
+                    capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> iSConnectWithWifi
+                    capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> iSConnectWithMobileData
+                    capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> iSConnectWithEthernet
+                    else -> iSConnectWithUnknown
+                }
+                println("Network status: ${getStatusString(status)} ($status)")
+                return status
+            }
+        } else {
+            @Suppress("DEPRECATION")
+            val networkInfo = connectivityManager.activeNetworkInfo
+            if (networkInfo != null && networkInfo.isConnected) {
+                val status = when (networkInfo.type) {
+                    ConnectivityManager.TYPE_WIFI -> iSConnectWithWifi
+                    ConnectivityManager.TYPE_MOBILE -> iSConnectWithMobileData
+                    ConnectivityManager.TYPE_ETHERNET -> iSConnectWithEthernet
+                    else -> iSConnectWithUnknown
+                }
+                println("Network status: ${getStatusString(status)} ($status)")
+                return status
+            }
+        }
+
+        println("Network status: no_connection")
+        return noInternetConnect
+    }
+
+    private fun getStatusString(status: Int): String {
+        return when (status) {
+            1 -> "iSConnectWithWifi"
+            2 -> "iSConnectWithMobileData"
+            3 -> "iSConnectWithEthernet"
+            0 -> "iSConnectWithUnknown"
+            else -> "Unknown status"
+        }
+    }
+
+    private fun getBatteryChargingStatus(): Int {
+        val batteryStatus: Intent? =
+            IntentFilter(Intent.ACTION_BATTERY_CHANGED).let { intentFilter ->
+                applicationContext.registerReceiver(null, intentFilter)
+            }
+
+        val charging = 0
+        val discharging = 1
+        val notCharging = 2
+
+        val status = batteryStatus?.getIntExtra(BatteryManager.EXTRA_STATUS, -1)
+        val isCharging =
+            status == BatteryManager.BATTERY_STATUS_CHARGING || status == BatteryManager.BATTERY_STATUS_FULL
+
+        val chargingStatus = when {
+            isCharging -> charging
+            status == BatteryManager.BATTERY_STATUS_DISCHARGING -> discharging
+            else -> notCharging
+        }
+
+        println("Charging status: ${getChargingStatusString(chargingStatus)} ($chargingStatus)")
+        return chargingStatus
+    }
+
+    private fun getChargingStatusString(status: Int): String {
+        return when (status) {
+            0 -> "charging"
+            1 -> "discharging"
+            2 -> "notCharging"
+            else -> "Unknown status"
+        }
+    }
+
+
 }